+++ R/app_bootstrap.R
@@ -12,6 +12,7 @@ required_packages <- c(
   "haven",
   "readr",
   "readxl",
+  "cellranger",
   "htmltools",
   "markdown",
   "openxlsx",